pyproject.toml
[project]
name = "source"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arch>=8.0.0",
    "ccxt>=4.5.24",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "prefect>=3.6.4",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "streamlit>=1.52.0",
    "xgboost>=3.1.2",
    "yfinance>=0.2.66",
]
